echo/requirements.txt
python-socketio == 5
python-engineio == 4
Requests
